Refactored version related entities into model subpackage.
